ASoC: SOF: Intel: move common ATOM stuff to module
Split between ACPI/PCI parts and use common module.

Since it's a split of existing code, the same dual-license is used for
the new atom.c and atom.h files.
